move type caster to a header file
